updated playsound into subfunctions/controlled behavior
