+++ playsound.py
@@ -40,40 +40,53 @@ def multi_sound(filenames):
         play_obj = wave_obj.play()
     play_obj.wait_done()
 
-
-
 def play_sound_func():
-    if sys.argv[1] == "-m" or sys.argv[1] == "--multi":
-        # tests
-        try:
-            float(sys.argv[2])              # if the first argument is a number...
-            sys.argv.remove(sys.argv[2])    # get rid of it as it will cause issues in the code.
-        except ValueError:
-            pass
+    # also need to have a default/work without an internal flag... need to see if 
+    # sys.argv[2] is a valid sound if it isn't one of our accepted flags.
+    if sys.argv[2] == "-m" or sys.argv[2] == "--multi":
+        play_multisound_func(3)
+        
+    elif sys.argv[2] == "-s" or sys.argv[2] == "--sequence":
+        play_seqsound_func(3)
 
-        # code to play the sounds
-        if sys.argv[2]:                                # if there is a sound to play... (sys.argv[2] not null)
-            for arg in sys.argv[2:]:                   # then append all of the sounds we have together...
-                sounds.append(arg)
-            print(f'Playing sounds')
-            multi_sound(sounds)                        # and play them by calling multi_sound with the appended sounds.
+def play_multisound_func(index):
+    """ Play multiple sounds at once.
+    Arguments:
+    the index the sounds start at (or a list of sounds?)
+    """
+    # tests
+    try:
+        float(sys.argv[index])              # if the first argument is a number...
+        sys.argv.remove(sys.argv[index])    # get rid of it as it will cause issues in the code.
+    except ValueError:
+        pass
 
-    elif sys.argv[1] == "-s" or sys.argv[1] == "--sequence":
-        if argvlen < 3:
-            print('No file provided...')
+    # code to play the sounds
+    if sys.argv[index]:                                # if there is a sound to play... (sys.argv[2] not null)
+        for arg in sys.argv[index:]:                   # then append all of the sounds we have together...
+            sounds.append(arg)
+        print(f'Playing sounds')
+        multi_sound(sounds)                        # and play them by calling multi_sound with the appended sounds.
 
-        # tests (and delay calculations)
-        try:
-            delay = float(sys.argv[2])                 # if argument casts to float without issue, save as the delay. 
-        except:
-            delay = None                               # if there's an error, then it's probably a string (should be a sound).
+def play_seqsound_func(index):
+    """ Play multiple sounds one after the other.
+    Arguments:
+    the index the sounds start at (or a list of sounds?)
+    """
+    if argvlen < index + 1:
+        print('No file provided...')
+
+    # tests (and delay calculations)
+    try:
+        delay = float(sys.argv[index])                 # if argument casts to float without issue, save as the delay. 
+    except:
+        delay = None                               # if there's an error, then it's probably a string (should be a sound).
             
-        # we have sound(s) to play, so the sound(s) is played (one after the other) and its name is printed when it does.
-        if argvlen >= 3 and delay == None:
-            for sound in sys.argv[2:]:                # loops through remaining sounds and plays them (works with 1 sound).
-                print(f'Playing {sound}')
-                play_sound(sound)
-        elif argvlen >= 4 and delay:                  # we need one more arg (because delay takes up a arg), but play sounds.
-            print(f"Adding {str(delay)}s of delay between sounds.")
-            delay_sound(sys.argv[3:], delay)          # functions more like -p, loop in the function rather than in call.
-
+    # we have sound(s) to play, so the sound(s) is played (one after the other) and its name is printed when it does.
+    if argvlen >= index+1 and delay == None:
+        for sound in sys.argv[index:]:                # loops through remaining sounds and plays them (works with 1 sound).
+            print(f'Playing {sound}')
+            play_sound(sound)
+    elif argvlen >= index+2 and delay:                  # we need one more arg (because delay takes up a arg), but play sounds.
+        print(f"Adding {str(delay)}s of delay between sounds.")
+        delay_sound(sys.argv[index+1:], delay)          # functions more like -p, loop in the function rather than in call.
